change: Move the lib folder in app folder
