+++ app/api/webhooks/clerk/route.ts
@@ -4,8 +4,7 @@ import { headers } from "next/headers";
 import { NextResponse } from "next/server";
 import { Webhook } from "svix";
 
-import { createUser } from "@/lib/actions/user.action";
-import { Z_DATA_ERROR } from "zlib";
+import { createUser } from "@/app/lib/actions/user.action";
 
 export async function POST(req: Request) {
   const WEBHOOK_SECRET = process.env.WEBHOOK_SECRET;
