--- prisma/schema.prisma
@@ -1,22 +1,0 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-generator client {
-  provider = "prisma-client-js"
-}
-
-datasource db {
-  provider = "cockroachdb"
-  url = env("DATABASE_URL")
-}
-
-
-model Haircuts {
-  id          String   @id @default(cuid())
-  createdAt   DateTime @default(now())
-  updatedAt   DateTime @updatedAt
-  name        String
-  price       Float
-  description String
-  active      Boolean  @default(true)
-}
